refactored codebase and added todo list
